+++ AFQ/viz/utils.py
@@ -335,25 +335,25 @@ class Viz:
         self.backend = backend
         if backend == "fury":
             try:
-                import AFQ.viz.afq_fury
+                import AFQ.viz.fury_backend
             except ImportError:
                 raise ImportError(viz_import_msg_error("fury"))
-            self.visualize_bundles = AFQ.viz.afq_fury.visualize_bundles
-            self.visualize_roi = AFQ.viz.afq_fury.visualize_roi
-            self.visualize_volume = AFQ.viz.afq_fury.visualize_volume
-            self.create_gif = AFQ.viz.afq_fury.create_gif
-            self.stop_creating_gifs = AFQ.viz.afq_fury.stop_creating_gifs
+            self.visualize_bundles = AFQ.viz.fury_backend.visualize_bundles
+            self.visualize_roi = AFQ.viz.fury_backend.visualize_roi
+            self.visualize_volume = AFQ.viz.fury_backend.visualize_volume
+            self.create_gif = AFQ.viz.fury_backend.create_gif
+            self.stop_creating_gifs = AFQ.viz.fury_backend.stop_creating_gifs
         elif backend == "plotly":
             try:
-                import AFQ.viz.afq_plotly
+                import AFQ.viz.plotly_backend
             except ImportError:
                 raise ImportError(viz_import_msg_error("plotly"))
-            self.visualize_bundles = AFQ.viz.afq_plotly.visualize_bundles
-            self.visualize_roi = AFQ.viz.afq_plotly.visualize_roi
-            self.visualize_volume = AFQ.viz.afq_plotly.visualize_volume
-            self.create_gif = AFQ.viz.afq_plotly.create_gif
+            self.visualize_bundles = AFQ.viz.plotly_backend.visualize_bundles
+            self.visualize_roi = AFQ.viz.plotly_backend.visualize_roi
+            self.visualize_volume = AFQ.viz.plotly_backend.visualize_volume
+            self.create_gif = AFQ.viz.plotly_backend.create_gif
             self.stop_creating_gifs = \
-                AFQ.viz.afq_plotly.stop_creating_gifs
+                AFQ.viz.plotly_backend.stop_creating_gifs
 
 
 def visualize_tract_profiles(tract_profiles, scalar="dti_fa", min_fa=0.0,
